Fix compiler warnings with -Wextra

+++ python_orocos_kdl/PyKDL/pybind11/dynamics.cpp
@@ -49,7 +49,7 @@ void init_dynamics(pybind11::module &m)
     {
         int i = std::get<0>(idx);
         int j = std::get<1>(idx);
-        if (i < 0 || i >= jm.rows() || j < 0 || j >= jm.columns())
+        if (i < 0 || (unsigned int)i >= jm.rows() || j < 0 || (unsigned int)j >= jm.columns())
             throw py::index_error("Inertia index out of range");
 
         return jm((unsigned int)i, (unsigned int)j);
@@ -58,7 +58,7 @@ void init_dynamics(pybind11::module &m)
     {
         int i = std::get<0>(idx);
         int j = std::get<1>(idx);
-        if (i < 0 || i >= jm.rows() || j < 0 || j >= jm.columns())
+        if (i < 0 || (unsigned int)i >= jm.rows() || j < 0 || (unsigned int)j >= jm.columns())
             throw py::index_error("Inertia index out of range");
 
         jm((unsigned int)i, (unsigned int)j) = value;
